+++ backend/bidding-service/src/main/java/com/mazad/bidding_service/domain/bid/BidValidator.java
@@ -1,7 +1,5 @@
 package com.mazad.bidding_service.domain.bid;
 
-import java.math.BigDecimal;
-
 import com.mazad.bidding_service.domain.auction.Auction;
 import com.mazad.bidding_service.domain.auction.AuctionStatus;
 import com.mazad.bidding_service.domain.exception.AuctionClosedException;
@@ -9,7 +7,7 @@ import com.mazad.bidding_service.domain.exception.InvalidBidAmountException;
 
 public class BidValidator {
 
-    public static void validate(Auction auction, BigDecimal amount) {
+    public static void validate(Auction auction, Long amount) {
         
         if (auction.getStatus() == AuctionStatus.CLOSED) {
             throw new AuctionClosedException();
