+++ backend/bidding-service/src/main/java/com/mazad/bidding_service/infrastructure/kafka/AuctionUpdateEvent.java
@@ -1,13 +1,12 @@
 package com.mazad.bidding_service.infrastructure.kafka;
 
-import java.math.BigDecimal;
 import java.time.Instant;
 
 import com.mazad.bidding_service.domain.auction.AuctionStatus;
 
 public record AuctionUpdateEvent(
         Long auctionId,
-        BigDecimal currentHighestBid,
+        Long currentHighestBid,
         Instant endsAt,
         AuctionStatus status,
         Long lastBidderId
